inventory_export_app/requirements.txt
Flask==3.1.0
requests==2.31.0
pandas==2.2.3
googletrans==4.0.0-rc1
python-dotenv==1.0.0
openpyxl==3.1.5
gunicorn==21.2.0